+++ InterlopingArtifact/InterRefs.cs
@@ -1,5 +1,6 @@
 using System.Diagnostics.CodeAnalysis;
 using RoR2;
+using UnityEngine;
 
 namespace HDeMods {
     [SuppressMessage("ReSharper", "InconsistentNaming")]
@@ -10,7 +11,12 @@ namespace HDeMods {
         internal const uint sfxBellToll = 2239523887;
         internal const uint sfxBellFinal = 1844540516;
         internal static BodyIndex FlyingVermin;
+        internal static EquipmentIndex VoidAspect = EquipmentIndex.None;
+        internal static readonly int remapTex = Shader.PropertyToID("_RemapTex");
 
-        internal static void CacheBlindPest() => FlyingVermin = BodyCatalog.FindBodyIndex("FlyingVerminBody");
+        internal static void CacheIndexes() {
+            FlyingVermin = BodyCatalog.FindBodyIndex("FlyingVerminBody");
+            VoidAspect = EquipmentCatalog.FindEquipmentIndex(DLC1Content.Elites.Void.eliteEquipmentDef.nameToken);
+        }
     }
 }