+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+- Added Respect Enemy Limit option.
+- Added Aggressive Culling option.
+- Added Culling Radius option.
+- These options are intended to both improve performance and keep gameplay fast paced by removing enemies that are not actively engaging with players.
+- Config options are now clamped.
 ## 0.1.3
 - Fixed a bug that caused resetting options to only work on the title screen
 ## 0.1.2